Crate fractal_utils

This crate contains utilities to work with Fractal Global Credits. It contains the wallet address representation and the currency amount type. Each of them is explained in its own documentation and can be easily used by third parties.

Using it is as simple as including this in the crate:

extern crate fractal_utils;

Reexports

pub use amount::Amount;

pub use wallet_address::{WALLET_ADDRESS_LEN, WalletAddress};

pub use location::Address;

Modules

amount
Fractal Global Credits amount

location
Fractal Global Location module

wallet_address
Fractal Global Wallet Address

Constants

CURRENCY_SYMBOL
The symbol of Fractal Global Credits