Module fractal_utils::amount

Fractal Global Credits amount

This module holds the Amount type and the AmountParseError. It will eventually hold MAX and MIN values for Amounts when constant expressions are implemented in the compiler in the stable chanel.

The maximum and minimum amount values can in any case be known by using max_value() and min_value() functions in the Amount type:

use std::u64;
use fractal_utils::Amount;

let max_value = Amount::max_value();
let min_value = Amount::min_value();

assert_eq!(max_value, Amount::from_repr(u64::MAX));
assert_eq!(min_value, Amount::from_repr(u64::MIN));

Structs

Amount
Fractal Global Credits amount

AmountParseError
Amount parsing error.