Struct fractal_utils::amount::Amount

pub struct Amount {
    // some fields omitted
}

Fractal Global Credits amount

This struct can be used the same way as any other number. An Amount can be added or substracted to another Amount, and it can be divided and multiplied by an integer. All operations that are defined in the Amount scope and that are exact can be used directly as usual integer / float point operations.

No negative amounts can exist, since an Amount is unsigned, sothe negation operator '-', then, has no use with an Amount.

Its internal representation is a 64 bit unsigned number, that is displayed as a fixed point, number of factor 1/1,000. This means that an internal representation of 1,000 would be an external amount of 1. The internal representation shouldn't be used except when serializing and deserializing the data, since this type is sent in JSON as its internal u64.

The use is the following:

use fractal_utils::Amount;

let amount = Amount::from_repr(1_654); // 1.654
let ten = Amount::from_repr(10_000); // 10
let add_ten = amount + ten;
assert_eq!(add_ten, Amount::from_repr(11_654)); // 11.654

They can be divided and multiplied by any other unsigned integer:

let mut amount = Amount::from_repr(7_000); // 7
amount *= 10u32;
assert_eq!(amount, Amount::from_repr(70_000)); // 70

amount = amount / 30u16;
assert_eq!(amount, Amount::from_repr(2_333)); // 2.333

amount %= 1u8;
assert_eq!(amount, Amount::from_repr(333)); // 0.333

Amounts can easily be displayed using the Display trait as any other number:

let amount = Amount::from_repr(56_000);
assert_eq!(format!("{}", amount), "56");
assert_eq!(format!("{:.2}", amount), "56.00");
assert_eq!(format!("{:.5}", amount), "56.00000");
assert_eq!(format!("{:05.1}", amount), "056.0");

// And with rounding:
let amount = Amount::from_repr(56); // 0.056
assert_eq!(format!("{:.2}", amount), "0.06");

Methods

impl Amount

fn from_repr(value: u64) -> Amount

Creates a new amount from its internal representation.

fn get_repr(&self) -> u64

Gets the internal representation of the amount.

fn min_value() -> Amount

Returns the smallest value that can be represented as a currency amount.

fn max_value() -> Amount

Returns the largest value that can be represented as a currency amount.